refactor: mudando o caminho das imagens
