+++ index.html
@@ -10,7 +10,7 @@
     <meta name="description" content="Plataforma Healz para check-in de pacientes e usuários." />
     <meta name="Equipe A" content="Healz Team" />
 
-    <!-- Favicon -->
+    <!-- Favicon  -->
     <link rel="icon" href="src/assets/favicon.ico" />
 
     <!-- Progressive Enhancement -->
