+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from '../../assets/logo.svg';
+import logo from '../../assets/whiteLogo.png';
 
 import { HeaderContainer, Logo, Title, RightSpace } from '../../styles/Header.style';
